Changes:   * Added a Formatter class hierarchy   * Implemented an Aligned Formatter that outputs left-aligned columns separated by spaces.   * Fixed a minor bug when executing a faulty query - prepared statement was not finalized sometimes - resource leak.   * Fixed a minor bug in the Flag class hierarchy - BoolFlag was not properly displaying default values.

+++ src/database.hpp
@@ -53,7 +53,7 @@ class ResultSet {
   public:
     const HeadersType headers;
     const DataType data;
-    const int rows, columns;
+    const size_t rows, columns;
 
   // DISALLOWED:
   private:
@@ -120,3 +120,4 @@ class DBObject {
 }  // namespace ash
 
 #endif  /* __ASH_DATABASE__ */
+
